feature/modelagem | implementa o restante das regras de negócio e limpa código boilerplate dos controllers de Postagem e Comentário

+++ app/Http/Controllers/ComentarioController.php
@@ -2,34 +2,25 @@
 
 namespace App\Http\Controllers;
 
+use App\User;
 use App\Models\Comentario;
+use App\Models\TransacaoLog;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Cache;
+use App\Notifications\ComentarioNaPostagem;
 
 class ComentarioController extends Controller
 {
     /**
-     * Display a listing of the resource.
+     * Rota do endpoint para criação de comentários em postagens
+     * -> controlado por throttle (o usuário pode comentar até 3x por minuto)
+     * -> controlado por permissão (comentar-postagem)
      *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
-    {
-        //
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
+     * Regras:
+     * - O usuário tem que ter saldo disponível caso queira comprar destaque
+     * - O usuário ou o dono da postagem tem que ser um assinante, caso contrário
+     * o comentário só é salvo com compra de destaque
      *
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
@@ -42,58 +33,116 @@ class ComentarioController extends Controller
         $input['login'] = $user->email;
         $input['assinante'] = $user->is_assinante;
 
+        $usuarioTemSaldo = $user->verificarSaldo($input['quantidade_moedas']);
+        $tentouComprarDestaque = $input['compra_destaque'] == true;
+
+        $donoDaPostagem = \App\Models\Post::find($input['post_id'])->autor;
+        $donoIsAssinante = $donoDaPostagem->is_assinante == true;
+        $userIsAssinante = $user->is_assinante == true;
+
+        $condicionalSemAssinantes = $donoIsAssinante == false && $userIsAssinante == false && $tentouComprarDestaque == false;
+
+        if ($condicionalSemAssinantes) {
+            return $this->sendError('O comentário não pode ser gravado pois o  dono da postagem e o usuário não são assinantes e o usuário não comprou destaque');
+        }
+
+        if ($tentouComprarDestaque && !$usuarioTemSaldo) {
+            return $this->sendError('O usuário não tem saldo para realizar esta compra.');
+        }
+
+        if ($tentouComprarDestaque && $usuarioTemSaldo) {
+            $this->comprarDestaque($user, $input);
+        }
+
+        $this->limparCacheDeComentariosDaPostagem($input['post_id']);
         $comentario = Comentario::create($input);
+
+        $postagem = $comentario->post;
+
+        $donoDaPostagem = $postagem->autor;
+        $donoDaPostagem->notify(new ComentarioNaPostagem($postagem, $user));
 
         return $this->sendResponse($comentario, 'Comentário criado com sucesso.');
     }
 
     /**
-     * Display the specified resource.
+     * Tenta comprar X moedas (1 moeda seria 1 real, pra facilitar)
+     * antes de salvar comentários com a flag compra_destaque no input
      *
-     * @param  \App\Models\Comentario  $comentario
-     * @return \Illuminate\Http\Response
+     * @return void || Error
      */
-    public function show(Comentario $comentario)
+    public function comprarDestaque($user, $comentario)
     {
-        //
+        $quantidadeDeMoedas = $comentario['quantidade_moedas'];
+
+        $user->debitar($quantidadeDeMoedas);
+
+        TransacaoLog::create([
+            'user_id' => $user->id,
+            'creditos' => 0,
+            'debitos' => $quantidadeDeMoedas,
+            'retencao_interna' => false,
+        ]);
+
+        $this->gerarRetencaoDoSistema($quantidadeDeMoedas);
     }
 
     /**
-     * Show the form for editing the specified resource.
+     * Método responsável por criar a transação de retenção interna
+     * de comentários comprando destaque para o sistema
      *
-     * @param  \App\Models\Comentario  $comentario
-     * @return \Illuminate\Http\Response
+     * @return void
      */
-    public function edit(Comentario $comentario)
+    public function gerarRetencaoDoSistema($quantidadeDeMoedas)
     {
-        //
+        $usuarioSistema = User::whereEmail('sistema')->first();
+
+        $porcentagemDeRetencaoParaComentarios = 0.05;
+        $valorRetido = $quantidadeDeMoedas * $porcentagemDeRetencaoParaComentarios;
+
+        TransacaoLog::create([
+            'user_id' => $usuarioSistema->id,
+            'creditos' => $valorRetido,
+            'debitos' => 0,
+            'retencao_interna' => true,
+        ]);
+
+        $usuarioSistema->creditar($valorRetido);
     }
 
     /**
-     * Update the specified resource in storage.
+     * Antes de salvar um comentário novo é necessário limpar o cache para que a listagem seja atualizada
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Comentario  $comentario
-     * @return \Illuminate\Http\Response
+     * @return void
      */
-    public function update(Request $request, Comentario $comentario)
+    public function limparCacheDeComentariosDaPostagem($id)
     {
-        //
+        if (Cache::has("postagem_$id")) {
+            Cache::forget("postagem_$id");
+        }
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Rota para exclusão de comentários
+     * Regra: só o dono do comentário ou da postagem relacionada podem excluir o comentário
      *
      * @param  \App\Models\Comentario  $comentario
      * @return \Illuminate\Http\Response
      */
     public function delete($id)
     {
-        //TODO: regras para exclusão
         $comentario = Comentario::find($id);
 
         if (!$comentario) {
             return $this->sendError('Comentário não encontrado');
+        }
+
+        $isDonoDoComentario = $comentario->user_id == Auth::id();
+        $isDonoDaPostagem = $comentario->post->user_id == Auth::id();
+        $naoPodeExcluir = $isDonoDoComentario == false && $isDonoDaPostagem == false;
+
+        if ($naoPodeExcluir) {
+            return $this->sendError('Você não tem permissão para executar essa ação');
         }
 
         $comentario->delete();
